Start reworking Server input side
