+++ src/Server.cpp
@@ -32,24 +32,11 @@ Server::Server()
   wl_signal_add(&xdg_shell->events.new_surface, &new_xdg_surface);
 
   cursor = new ServerCursor(this);
-  // cursor = wlr_cursor_create();
-  // wlr_cursor_attach_output_layout(cursor, output_layout);
-  //
-  // cursor_mgr = wlr_xcursor_manager_create(nullptr, 24);
-  // wlr_xcursor_manager_load(cursor_mgr, 1);
-  //
-  // cursor_motion.notify = ServerCursor::server_cursor_motion;
-  // wl_signal_add(&cursor->events.motion, &cursor_motion);
-  // cursor_motion_absolute.notify = ServerCursor::server_cursor_motion_absolute;
-  // wl_signal_add(&cursor->events.motion_absolute, &cursor_motion_absolute);
-  // cursor_button.notify = ServerCursor::server_cursor_button;
-  // wl_signal_add(&cursor->events.button, &cursor_button);
-  // cursor_axis.notify = ServerCursor::server_cursor_axis;
-  // wl_signal_add(&cursor->events.axis, &cursor_axis);
+  input = new ServerInput(this);
+  // wl_list_init(&keyboards);
+  // new_input.notify = ServerInput::server_new_input;
+  // wl_signal_add(&backend->events.new_input, &new_input);
 
-  wl_list_init(&keyboards);
-  new_input.notify = ServerInput::server_new_input;
-  wl_signal_add(&backend->events.new_input, &new_input);
   seat = wlr_seat_create(display, "seat0");
   request_cursor.notify = Seat::seat_request_cursor;
   wl_signal_add(&seat->events.request_set_cursor, &request_cursor);
