+++ src/main.cpp
@@ -1,6 +1,6 @@
 #include "Server.hpp"
 
-int main(int argc, char **argv)
+int main()
 {
   Server server;
